+++ binance/spot/gift_card.py
@@ -41,7 +41,8 @@ def gift_card_redeem_code(self, code: str, **kwargs):
     Args:
       code (str): Binance Code
     Keyword Args:
-      recvWindow (int, optional): The value cannot be greater than 60000
+        externalUid (str, optional): Each external unique ID represents a unique user on the partner platform. The function helps you to identify the redemption behavior of different users, such as redemption frequency and amount. It also helps risk and limit control of a single account, such as daily limit on redemption volume, frequency, and incorrect number of entries. This will also prevent a single user account reach the partner's daily redemption limits. We strongly recommend you to use this feature and transfer us the User ID of your users if you have different users redeeming Binance codes on your platform. To protect user data privacy, you may choose to transfer the user id in any desired format (max. 400 characters).
+        recvWindow (int, optional): The value cannot be greater than 60000
     """
     check_required_parameter(code, "code")
 
@@ -71,3 +72,24 @@ def gift_card_verify_code(self, referenceNo: str, **kwargs):
     payload = {"referenceNo": referenceNo, **kwargs}
 
     return self.sign_request("GET", "/sapi/v1/giftcard/verify", payload)
+
+
+def gift_card_rsa_public_key(self, **kwargs):
+    """Fetch RSA Public Key (USER_DATA)
+
+    This API is for fetching the RSA Public Key.
+    This RSA Public key will be used to encrypt the card code.
+    Please note that the RSA Public key fetched is valid only for the current day.
+
+    Weight(IP): 1
+
+    GET /sapi/v1/giftcard/cryptography/rsa-public-key
+
+    https://binance-docs.github.io/apidocs/spot/en/#fetch-rsa-public-key-user_data
+
+    Keyword Args:
+        recvWindow (int, optional): The value cannot be greater than 60000
+    """
+
+    url_path = "/sapi/v1/giftcard/cryptography/rsa-public-key"
+    return self.sign_request("GET", url_path, {**kwargs})