+++ examples/spot/market/book_ticker.py
@@ -9,3 +9,4 @@ config_logging(logging, logging.DEBUG)
 spot_client = Client(base_url="https://testnet.binance.vision")
 
 logging.info(spot_client.book_ticker("BTCUSDT"))
+logging.info(spot_client.book_ticker(symbols=["BTCUSDT", "BNBUSDT"]))