+++ examples/spot/staking/staking_set_auto_staking.py
@@ -1,0 +1,25 @@
+#!/usr/bin/env python
+
+import logging
+from binance.spot import Spot as Client
+from binance.lib.utils import config_logging
+from binance.error import ClientError
+
+config_logging(logging, logging.DEBUG)
+
+key = ""
+secret = ""
+
+params = {"product": "STAKING", "positionId": "1234", "renewable": "true"}
+
+client = Client(key, secret)
+
+try:
+    response = client.staking_set_auto_staking(**params)
+    logging.info(response)
+except ClientError as error:
+    logging.error(
+        "Found error. status: {}, error code: {}, error message: {}".format(
+            error.status_code, error.error_code, error.error_message
+        )
+    )